Remove pre-post brexit question for manual flow (#2037)
Default to "post brexit" behaviour enforcing the full manual flow.
Previously submitted notifications will still have the
"was_notified_before_eu_exit" flag in DB as "true", but new
notifications will rely on the default "false" value.

The question has been removed from the manual flow.
