+++ cosmetics-web/app/controllers/responsible_persons/notifications_controller.rb
@@ -39,12 +39,7 @@ class ResponsiblePersons::NotificationsController < SubmitApplicationController
   def show; end
 
   def new
-    was_notified_before_eu_exit = params["notified_before_eu_exit"] == "true"
-
-    @notification = Notification.create(
-      responsible_person: @responsible_person,
-      was_notified_before_eu_exit: was_notified_before_eu_exit,
-    )
+    @notification = Notification.create(responsible_person: @responsible_person)
 
     redirect_to new_responsible_person_notification_build_path(@responsible_person, @notification)
   end
