+++ cosmetics-web/app/views/responsible_persons/wizard/add_notification_wizard/register_on_eu_system.html.erb
@@ -23,8 +23,8 @@
       "https://ec.europa.eu/growth/sectors/cosmetics/cpnp_en", class: 'govuk-link' %></p>
 
     <p><%= link_to "Notify OPSS without using a ZIP file",
-        new_responsible_person_notification_path(@responsible_person,
-        notified_before_eu_exit: false), class: 'govuk-link' %></p>
+                   new_responsible_person_notification_path(@responsible_person),
+                   class: 'govuk-link' %></p>
 
   </div>
 </div>
