--- cosmetics-web/app/views/responsible_persons/wizard/add_notification_wizard/was_product_on_sale_before_eu_exit.html.erb
@@ -1,22 +1,0 @@
-<% eu_exit_date = display_full_month_date(EU_EXIT_DATE) %>
-<% title_without_date = "Was this product notified in the EU before " %>
-<% content_for :page_title, title_without_date + eu_exit_date + '?' %>
-<% content_for :after_header do %>
-  <%= link_to "Back", wizard_path(:do_you_have_files_from_eu_notification), class: "govuk-back-link" %>
-<% end %>
-
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-    <%= form_with url: wizard_path, method: :put do |form| %>
-
-      <% question = safe_join([title_without_date, content_tag('span', eu_exit_date, class: 'no-wrap'), '?']) %>
-      <%= render "components/govuk_radios",
-          name: :answer,
-          errorMessage: @error_text.present? && { text: @error_text },
-          fieldset: { legend: { text: question, isPageHeading: true, classes: "govuk-fieldset__legend--l" } },
-          items: [{ text: "Yes", value: "yes" },
-            { text: "No", value: "no" }] %>
-      <%= govukButton text: "Continue" %>
-    <% end %>
-  </div>
-</div>
