+++ cosmetics-web/spec/features/manual_notifications_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe "Manual, pre-Brexit notifications", :with_stubbed_antivirus, type: :feature do
+RSpec.describe "Manual notifications", :with_stubbed_antivirus, type: :feature do
   let(:responsible_person) { create(:responsible_person_with_user, :with_a_contact_person) }
   let(:user) { responsible_person.responsible_person_users.first.user }
 
@@ -8,7 +8,7 @@ RSpec.describe "Manual, pre-Brexit notifications", :with_stubbed_antivirus, type
     sign_in_as_member_of_responsible_person(responsible_person, user)
   end
 
-  scenario "Manual, post-Brexit, exact ingredients, single item, with CMRS, no nanomaterials" do
+  scenario "Manual, exact ingredients, single item, with CMRS, no nanomaterials" do
     visit new_responsible_person_add_notification_path(responsible_person)
 
     expect_to_be_on__was_eu_notified_about_products_page
@@ -35,7 +35,7 @@ RSpec.describe "Manual, pre-Brexit notifications", :with_stubbed_antivirus, type
     expect_back_link_to_is_product_for_under_threes_page
     answer_is_product_multi_item_kit_with "No, this is a single product"
 
-    exepct_to_be_on_upload_product_label_page
+    expect_to_be_on_upload_product_label_page
     expect_back_link_to_multi_item_kits_page
     upload_product_label
 
@@ -114,7 +114,7 @@ RSpec.describe "Manual, pre-Brexit notifications", :with_stubbed_antivirus, type
     expect_to_see_message "SkinSoft tangerine shampoo notification submitted"
   end
 
-  scenario "Manual, post-Brexit, exact ingredients, single item, no nanomaterials" do
+  scenario "Manual, exact ingredients, single item, no nanomaterials" do
     visit new_responsible_person_add_notification_path(responsible_person)
 
     expect_to_be_on__was_eu_notified_about_products_page
@@ -141,7 +141,7 @@ RSpec.describe "Manual, pre-Brexit notifications", :with_stubbed_antivirus, type
     expect_back_link_to_is_product_for_under_threes_page
     answer_is_product_multi_item_kit_with "No, this is a single product"
 
-    exepct_to_be_on_upload_product_label_page
+    expect_to_be_on_upload_product_label_page
     expect_back_link_to_multi_item_kits_page
     upload_product_label
 
@@ -214,7 +214,7 @@ RSpec.describe "Manual, pre-Brexit notifications", :with_stubbed_antivirus, type
     expect_to_see_message "SkinSoft tangerine shampoo notification submitted"
   end
 
-  scenario "Manual, post-Brexit, ingredient ranges, single item, no nanomaterials" do
+  scenario "Manual, ingredient ranges, single item, no nanomaterials" do
     visit new_responsible_person_add_notification_path(responsible_person)
 
     expect_to_be_on__was_eu_notified_about_products_page
@@ -241,7 +241,7 @@ RSpec.describe "Manual, pre-Brexit notifications", :with_stubbed_antivirus, type
     expect_back_link_to_is_product_for_under_threes_page
     answer_is_product_multi_item_kit_with "No, this is a single product"
 
-    exepct_to_be_on_upload_product_label_page
+    expect_to_be_on_upload_product_label_page
     expect_back_link_to_multi_item_kits_page
     upload_product_label
 
@@ -314,7 +314,7 @@ RSpec.describe "Manual, pre-Brexit notifications", :with_stubbed_antivirus, type
     expect_to_see_message "SkinSoft tangerine shampoo notification submitted"
   end
 
-  scenario "Manual, post-Brexit, frame formulation, single item, no nanomaterials" do
+  scenario "Manual, frame formulation, single item, no nanomaterials" do
     visit new_responsible_person_add_notification_path(responsible_person)
 
     expect_to_be_on__was_eu_notified_about_products_page
@@ -341,7 +341,7 @@ RSpec.describe "Manual, pre-Brexit notifications", :with_stubbed_antivirus, type
     expect_back_link_to_is_product_for_under_threes_page
     answer_is_product_multi_item_kit_with "No, this is a single product"
 
-    exepct_to_be_on_upload_product_label_page
+    expect_to_be_on_upload_product_label_page
     expect_back_link_to_multi_item_kits_page
     upload_product_label
 
@@ -418,7 +418,7 @@ RSpec.describe "Manual, pre-Brexit notifications", :with_stubbed_antivirus, type
     expect_to_see_message "SkinSoft tangerine shampoo notification submitted"
   end
 
-  scenario "Manual, post-Brexit, frame formulation (with poisonous ingredients), single item, no nanomaterials" do
+  scenario "Manual, frame formulation (with poisonous ingredients), single item, no nanomaterials" do
     visit new_responsible_person_add_notification_path(responsible_person)
 
     expect_to_be_on__was_eu_notified_about_products_page
@@ -445,7 +445,7 @@ RSpec.describe "Manual, pre-Brexit notifications", :with_stubbed_antivirus, type
     expect_back_link_to_is_product_for_under_threes_page
     answer_is_product_multi_item_kit_with "No, this is a single product"
 
-    exepct_to_be_on_upload_product_label_page
+    expect_to_be_on_upload_product_label_page
     expect_back_link_to_multi_item_kits_page
     upload_product_label
 
@@ -526,5 +526,210 @@ RSpec.describe "Manual, pre-Brexit notifications", :with_stubbed_antivirus, type
 
     expect_to_be_on__your_cosmetic_products_page
     expect_to_see_message "SkinSoft tangerine shampoo notification submitted"
+  end
+
+  scenario "Manual, frame formulation, multi-item, no nanomaterials, no poison" do
+    visit new_responsible_person_add_notification_path(responsible_person)
+
+    expect_to_be_on__was_eu_notified_about_products_page
+    expect_back_link_to_notifications_page
+    answer_was_eu_notified_with "Yes"
+
+    expect_to_be_on__do_you_have_the_zip_files_page
+    expect_back_link_to_was_eu_notified_about_products_page
+    answer_do_you_have_zip_files_with "No, I’ll enter information manually"
+
+    expect_to_be_on__what_is_product_called_page
+    expect_back_link_to_do_you_have_the_zip_files_page
+    answer_product_name_with "SkinSoft strawberry blonde hair dye"
+
+    expect_to_be_on__internal_reference_page
+    expect_back_link_to_what_is_product_called_page
+    answer_do_you_want_to_give_an_internal_reference_with "No"
+
+    expect_to_be_on__is_product_for_under_threes_page
+    expect_back_link_to_internal_reference_page
+    answer_is_product_for_under_threes_with "No"
+
+    expect_to_be_on__multi_item_kits_page
+    expect_back_link_to_is_product_for_under_threes_page
+    answer_is_product_multi_item_kit_with "Yes"
+
+    expect_to_be_on_upload_item_label_page
+    expect_back_link_to_multi_item_kits_page
+    upload_product_label
+
+    expect_to_be_on__how_are_items_used_together_page
+    expect_back_link_to_upload_item_label_page
+    answer_does_contain_items_that_need_to_be_mixed_with "No, the items are used in sequence"
+
+    expect_to_be_on__kit_items_page
+    expect_back_link_to_how_are_items_used_together_page
+    add_an_item
+
+    expect_to_be_on__what_is_item_called_page
+    expect_back_link_to_kit_items_page
+    answer_item_name_with "SkinSoft strawberry blonde hair colourant"
+
+    expect_to_be_on__is_item_available_in_shades_page(item_name: "SkinSoft strawberry blonde hair colourant")
+    expect_back_link_to_what_is_item_called_page
+    answer_is_item_available_in_shades_with "No", item_name: "SkinSoft strawberry blonde hair colourant"
+
+    expect_to_be_on__physical_form_of_item_page(item_name: "SkinSoft strawberry blonde hair colourant")
+    expect_back_link_to_is_item_available_in_shades_page
+    answer_what_is_physical_form_of_item_with "Liquid", item_name: "SkinSoft strawberry blonde hair colourant"
+
+    expect_to_be_on__what_is_product_contained_in_page(item_name: "SkinSoft strawberry blonde hair colourant")
+    expect_back_link_to_physical_form_of_item_page
+    answer_what_is_product_contained_in_with("A pressurised container, an impregnated sponge, wipe, patch or pad, or is encapsulated",
+                                             item_name: "SkinSoft strawberry blonde hair colourant")
+
+    expect_to_be_on__what_type_of_applicator_page
+    expect_back_link_to_what_is_product_contained_in_page
+    answer_what_type_of_applicator_with "Pressurised spray"
+
+    expect_to_be_on__does_item_contain_cmrs_page
+    expect_back_link_to_what_type_of_applicator_page
+    answer_does_item_contain_cmrs_with "No", item_name: "SkinSoft strawberry blonde hair colourant"
+
+    expect_to_be_on__does_item_contain_nanomaterial_page
+    expect_back_link_to_does_item_contain_cmrs_page
+    answer_does_item_contain_nanomaterials_with "No", item_name: "SkinSoft strawberry blonde hair colourant"
+
+    expect_to_be_on__item_category_page
+    expect_back_link_to_does_item_contain_nanomaterial_page
+    answer_item_category_with "Hair and scalp products"
+
+    expect_to_be_on__item_subcategoy_page(category: "hair and scalp products", item_name: "SkinSoft strawberry blonde hair colourant")
+    expect_back_link_to_item_category_page
+    answer_item_subcategory_with "Hair colouring products"
+
+    expect_to_be_on__item_sub_subcategory_page(subcategory: "hair colouring products", item_name: "SkinSoft strawberry blonde hair colourant")
+    expect_back_link_to_item_category_page("hair_and_scalp_products")
+    answer_item_sub_subcategory_with "Oxidative hair colour products"
+
+    expect_to_be_on__formulation_method_page item_name: "SkinSoft strawberry blonde hair colourant"
+    expect_back_link_to_item_category_page("hair_colouring_products")
+    answer_how_do_you_want_to_give_formulation_with("Choose a predefined frame formulation",
+                                                    item_name: "SkinSoft strawberry blonde hair colourant")
+
+    expect_to_be_on__frame_formulation_select_page
+    expect_back_link_to_formulation_method_page
+    give_frame_formulation_as "Hair Colorant (Permanent, Oxidative Type) - Type 1 : Two Components - Colorant Part"
+
+    expect_to_be_on__poisonous_ingredients_page
+    expect_back_link_to_frame_formulation_select_page
+    answer_does_product_contain_poisonous_ingredients_with "No"
+
+    expect_to_be_on__what_is_ph_range_of_product_page
+    expect_back_link_to_poisonous_ingredients_page
+    answer_what_is_ph_range_of_product_with "It does not have a pH"
+
+    expect_to_be_on__kit_items_page
+    expect_back_link_to_what_is_ph_range_of_product_page
+    add_an_item
+
+    expect_to_be_on__what_is_item_called_page
+    expect_back_link_to_kit_items_page
+    answer_item_name_with "SkinSoft strawberry blonde hair fixer"
+
+    expect_to_be_on__is_item_available_in_shades_page(item_name: "SkinSoft strawberry blonde hair fixer")
+    expect_back_link_to_what_is_item_called_page
+    answer_is_item_available_in_shades_with "No", item_name: "SkinSoft strawberry blonde hair fixer"
+
+    expect_to_be_on__physical_form_of_item_page item_name: "SkinSoft strawberry blonde hair fixer"
+    expect_back_link_to_is_item_available_in_shades_page
+    answer_what_is_physical_form_of_item_with "Liquid", item_name: "SkinSoft strawberry blonde hair fixer"
+
+    expect_to_be_on__what_is_product_contained_in_page(item_name: "SkinSoft strawberry blonde hair fixer")
+    expect_back_link_to_physical_form_of_item_page
+    answer_what_is_product_contained_in_with("A pressurised container, an impregnated sponge, wipe, patch or pad, or is encapsulated",
+                                             item_name: "SkinSoft strawberry blonde hair fixer")
+
+    expect_to_be_on__what_type_of_applicator_page
+    expect_back_link_to_what_is_product_contained_in_page
+    answer_what_type_of_applicator_with "Pressurised spray"
+
+    expect_to_be_on__does_item_contain_cmrs_page
+    expect_back_link_to_what_type_of_applicator_page
+    answer_does_item_contain_cmrs_with "No", item_name: "SkinSoft strawberry blonde hair fixer"
+
+    expect_to_be_on__does_item_contain_nanomaterial_page
+    expect_back_link_to_does_item_contain_cmrs_page
+    answer_does_item_contain_nanomaterials_with "No", item_name: "SkinSoft strawberry blonde hair fixer"
+
+    expect_to_be_on__item_category_page
+    expect_back_link_to_does_item_contain_nanomaterial_page
+    answer_item_category_with "Hair and scalp products"
+
+    expect_to_be_on__item_subcategoy_page(category: "hair and scalp products", item_name: "SkinSoft strawberry blonde hair fixer")
+    expect_back_link_to_item_category_page
+    answer_item_subcategory_with "Hair colouring products"
+
+    expect_to_be_on__item_sub_subcategory_page(subcategory: "hair colouring products", item_name: "SkinSoft strawberry blonde hair fixer")
+    expect_back_link_to_item_category_page("hair_and_scalp_products")
+    answer_item_sub_subcategory_with "Oxidative hair colour products"
+
+    expect_to_be_on__formulation_method_page item_name: "SkinSoft strawberry blonde hair fixer"
+    expect_back_link_to_item_category_page("hair_colouring_products")
+    answer_how_do_you_want_to_give_formulation_with("Choose a predefined frame formulation",
+                                                    item_name: "SkinSoft strawberry blonde hair fixer")
+
+    expect_to_be_on__frame_formulation_select_page
+    expect_back_link_to_formulation_method_page
+    give_frame_formulation_as "Hair / Scalp Care Lotion"
+
+    expect_to_be_on__poisonous_ingredients_page
+    expect_back_link_to_frame_formulation_select_page
+    answer_does_product_contain_poisonous_ingredients_with "No"
+
+    expect_to_be_on__what_is_ph_range_of_product_page
+    expect_back_link_to_poisonous_ingredients_page
+    answer_what_is_ph_range_of_product_with "It does not have a pH"
+
+    expect_to_be_on__kit_items_page
+    expect_back_link_to_what_is_ph_range_of_product_page
+    click_button "Continue"
+
+    expect_to_be_on__check_your_answers_page(product_name: "SkinSoft strawberry blonde hair dye")
+    expect_back_link_to_kit_items_page
+
+    expect_check_your_answers_page_for_kit_items_to_contain(
+      product_name: "SkinSoft strawberry blonde hair dye",
+      number_of_components: "2",
+      components_mixed: "No",
+      kit_items: [
+        {
+          name: "SkinSoft strawberry blonde hair colourant",
+          shades: "None",
+
+          nanomaterials: "None",
+          category: "Hair and scalp products",
+          subcategory: "Hair colouring products",
+          sub_subcategory: "Oxidative hair colour products",
+          formulation_given_as: "Frame formulation",
+          physical_form: "Liquid",
+          ph: "No pH",
+          poisonous_ingredients: "No",
+        },
+        {
+          name: "SkinSoft strawberry blonde hair fixer",
+          shades: "None",
+
+          nanomaterials: "None",
+          category: "Hair and scalp products",
+          subcategory: "Hair colouring products",
+          sub_subcategory: "Oxidative hair colour products",
+          formulation_given_as: "Frame formulation",
+          physical_form: "Liquid",
+          ph: "No pH",
+          poisonous_ingredients: "No",
+        },
+      ],
+    )
+    click_button "Accept and submit"
+
+    expect_to_be_on__your_cosmetic_products_page
+    expect_to_see_message "SkinSoft strawberry blonde hair dye notification submitted"
   end
 end
